# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Matka Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface Admin {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  admin: Admin;
}

export interface SignupRequest {
  name: string;
  phone: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  walletBalance: number;
  isBlocked: boolean;
  createdAt: string;
}

export interface SignupResponse {
  token: string;
  user: User;
}

export interface Bid {
  id: number;
  userId: number;
  userName?: string;
  marketId: number;
  marketName: string;
  gameType: string;
  amount: number;
  digit: string;
  openTime?: string;
  closeTime?: string;
  currentTime?: string;
  status: string;
  createdAt: string;
}

export interface DashboardStats {
  totalUsers: number;
  totalBidsToday: number;
  totalProfit: number;
  activeMarkets: number;
  depositsToday: number;
  withdrawalsToday: number;
  recentBids: Bid[];
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface UpdateUserRequest {
  isBlocked?: boolean;
  walletBalance?: number;
  name?: string;
  phone?: string;
}

export interface Market {
  id: number;
  name: string;
  openTime: string;
  closeTime: string;
  isActive: boolean;
  openResult?: string | null;
  closeResult?: string | null;
  jodiResult?: string | null;
  autoUpdate: boolean;
  sourceUrl?: string | null;
  lastFetchedAt?: string | null;
  fetchError?: string | null;
  createdAt: string;
}

export interface MarketAutoConfigRequest {
  autoUpdate: boolean;
  sourceUrl?: string | null;
}

export interface FetchResultResponse {
  success: boolean;
  message: string;
  openResult?: string | null;
  closeResult?: string | null;
  jodiResult?: string | null;
}

export interface ScraperLog {
  id: number;
  marketId?: number | null;
  marketName?: string | null;
  sourceUrl?: string | null;
  success: boolean;
  openResult?: string | null;
  closeResult?: string | null;
  jodiResult?: string | null;
  errorMessage?: string | null;
  createdAt: string;
}

export interface ScraperStatus {
  isRunning: boolean;
  totalMarketsWithAutoUpdate: number;
  lastRunAt?: string | null;
  nextRunIn: string;
}

export interface CreateMarketRequest {
  name: string;
  openTime: string;
  closeTime: string;
  isActive: boolean;
}

export interface Result {
  id: number;
  marketId: number;
  marketName: string;
  resultDate: string;
  openResult?: string | null;
  closeResult?: string | null;
  jodiResult?: string | null;
  pannaResult?: string | null;
  declaredAt?: string | null;
}

export interface DeclareResultRequest {
  marketId: number;
  resultDate: string;
  openResult?: string;
  closeResult?: string;
  jodiResult?: string;
  pannaResult?: string;
}

export interface BidListResponse {
  bids: Bid[];
  total: number;
  page: number;
  limit: number;
}

export interface Deposit {
  id: number;
  userId: number;
  userName: string;
  amount: number;
  status: string;
  paymentMethod: string;
  transactionId?: string | null;
  screenshotUrl?: string | null;
  createdAt: string;
  processedAt?: string | null;
}

export interface Withdrawal {
  id: number;
  userId: number;
  userName: string;
  amount: number;
  status: string;
  bankName?: string | null;
  accountNumber?: string | null;
  ifscCode?: string | null;
  upiId?: string | null;
  createdAt: string;
  processedAt?: string | null;
}

export interface GameRates {
  singleDigit: number;
  jodiDigit: number;
  singlePanna: number;
  doublePanna: number;
  triplePanna: number;
  halfSangam: number;
  fullSangam: number;
}

export interface Notice {
  id: number;
  title: string;
  content: string;
  isActive: boolean;
  createdAt: string;
}

export interface CreateNoticeRequest {
  title: string;
  content: string;
  isActive: boolean;
}

export interface AppSettings {
  appName: string;
  logoUrl?: string | null;
  supportPhone?: string | null;
  upiId?: string | null;
  bankName?: string | null;
  bankAccountNumber?: string | null;
  bankIfscCode?: string | null;
  qrCodeUrl?: string | null;
}

export type GetUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type GetScraperLogsParams = {
  limit?: number;
  marketId?: number;
};

export type GetResultsParams = {
  marketId?: number;
  date?: string;
};

export type GetBidsParams = {
  page?: number;
  limit?: number;
  marketId?: number;
  userId?: number;
  status?: string;
};

export type GetDepositsParams = {
  status?: string;
};

export type GetWithdrawalsParams = {
  status?: string;
};