# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Matka Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
...
  /deposits:
    get:
      operationId: getDeposits
      tags: [deposits]
      summary: List deposit requests
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
  /deposits/{id}/approve:
    post:
      operationId: approveDeposit
      tags: [deposits]
      summary: Approve a deposit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /deposits/{id}/reject:
    post:
      operationId: rejectDeposit
      tags: [deposits]
      summary: Reject a deposit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
  /notices:
    get:
      operationId: getNotices
      tags: [notices]
      summary: List notices
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notice"
    post:
      operationId: createNotice
      tags: [notices]
      summary: Create a notice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoticeRequest"
      responses:
        "201":
          description: Created notice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notice"
  /notices/{id}:
    delete:
      operationId: deleteNotice
      tags: [notices]
      summary: Delete a notice
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    DashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalBidsToday:
          type: integer
        totalProfit:
          type: number
        activeMarkets:
          type: integer
        depositsToday:
          type: number
        withdrawalsToday:
          type: number
        recentBids:
          type: array
          items:
            $ref: "#/components/schemas/Bid"
      required:
        - totalUsers
        - totalBidsToday
        - totalProfit
        - activeMarkets
        - depositsToday
        - withdrawalsToday
        - recentBids
...
    Bid:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        marketId:
          type: integer
        marketName:
          type: string
        gameType:
          type: string
        amount:
          type: number
        digit:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        currentTime:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - marketId
        - marketName
        - gameType
        - amount
        - digit
        - status
        - createdAt
...
    Deposit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        status:
          type: string
        paymentMethod:
          type: string
        transactionId:
          type: string
          nullable: true
        screenshotUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        processedAt:
          type: string
          nullable: true
      required:
        - id
        - userId
        - userName
        - amount
        - status
        - paymentMethod
        - createdAt
...
    Notice:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - content
        - isActive
        - createdAt
    CreateNoticeRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
      required:
        - title
        - content
        - isActive
...